SetupBehaviour

+++ TP_AI_Project/Assets/Teams/JabberMonkey/Behavior/Conditional/EnergyInfValue.cs
@@ -8,7 +8,7 @@ namespace JabberMonkey
 	[TaskCategory("JabberAI")]
 	public class EnergyInfValue : Conditional
 	{
-		public int energyValue;
+		public float energyValue;
 
 		private BlackBordScipt blackBord;
